tests.bash: redirect grep output to /dev/null
The man-page of grep says:

	-q, --quiet, --silent

	Quiet; do not write anything to standard output. Exit
	immediately with zero status if any match is found, even if an
	error was detected. Also see the -s or --no-messages option.

The grep option --quiet exits immediately according to its manual if a
pattern is found. This has an incidence if grep is used in a pipe as it
can exit before the left element(s) of the pipe.

In the situation of the pattern is found, the left side of the pipe
exits with return code 141 (SIGPIPE) because the right side exits first
(i.e. grep).

Therefore, it is preferable to redirect the output of grep to /dev/null
instead of using the option --quiet if grep is used in a pipe; this
leaves grep exit after the left element(s) of the pipeline exit.

Note: the pipeline fails if the bash option pipefail is set.

This commit removes the use of the grep option --quiet and redirects the
output to /dev/null.